LPS-101966 Don't create composite class names if they don't exist when deleting fragment entry link

+++ modules/apps/layout/layout-content-page-editor-web/src/main/java/com/liferay/layout/content/page/editor/web/internal/model/listener/FragmentEntryLinkModelListener.java
@@ -29,8 +29,10 @@ import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.model.BaseModelListener;
+import com.liferay.portal.kernel.model.ClassName;
 import com.liferay.portal.kernel.model.ModelListener;
 import com.liferay.portal.kernel.security.permission.ResourceActionsUtil;
+import com.liferay.portal.kernel.service.ClassNameLocalService;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
 import com.liferay.portal.kernel.util.Portal;
@@ -113,8 +115,15 @@ public class FragmentEntryLinkModelListener
 					ResourceActionsUtil.getCompositeModelName(
 						FragmentEntryLink.class.getName(), editableKey);
 
+				ClassName className = _classNameLocalService.fetchClassName(
+					compositeClassName);
+
+				if (className == null) {
+					continue;
+				}
+
 				_ddmTemplateLinkLocalService.deleteTemplateLink(
-					_portal.getClassNameId(compositeClassName),
+					className.getClassNameId(),
 					fragmentEntryLink.getFragmentEntryLinkId());
 			}
 		}
@@ -241,6 +250,9 @@ public class FragmentEntryLinkModelListener
 
 	@Reference
 	private AssetEntryUsageLocalService _assetEntryUsageLocalService;
+
+	@Reference
+	private ClassNameLocalService _classNameLocalService;
 
 	@Reference
 	private CommentManager _commentManager;
